# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ottoman Controller API
  version: 1.0.0
paths:
...
  /api/status/agent:
    get:
      summary: Get agent status
      operationId: getAgentStatus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/check:
    get:
      summary: Check authentication status
      operationId: checkAuth
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                    description: Whether the user is authenticated.
                    example: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/monitors:
    get:
      summary: Get all monitors
      operationId: getMonitors
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitorsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/layouts:
    get:
      summary: Get all layouts
      operationId: getLayouts
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/layouts/switch:
    post:
      summary: Switch layout
      operationId: switchLayout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchLayoutRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchLayoutResponse'
        '400':
          description: Bad Request (Invalid body or ambiguous layout)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/layouts/current:
    get:
      summary: Get current layout
      operationId: getCurrentLayout
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchLayoutResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/layouts/save-current:
    post:
      summary: Save current layout
      operationId: saveCurrentLayout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLayoutRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveLayoutResponse'
        '400':
          description: Bad Request (Invalid body or missing name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/layouts/remove:
    post:
      summary: Remove layout
      operationId: removeLayout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveLayoutRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveLayoutResponse'
        '400':
          description: Bad Request (Invalid body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Layout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/shutdown:
    post:
      summary: Shutdown agent
      operationId: shutdown
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShutdownResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway (Agent unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/sim/reset:
    post:
      summary: Reset simulation
      operationId: simReset
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimStateResponse'
        '404':
          description: Not Found (Controller not in simulation mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sim/state:
    get:
      summary: Get simulation state
      operationId: simState
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimStateResponse'
        '404':
          description: Not Found (Controller not in simulation mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sim/set-state:
    post:
      summary: Set simulation state
      operationId: simSetState
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimSetStateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimStateResponse'
        '404':
          description: Not Found (Controller not in simulation mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # NOTE: This endpoint upgrades to WebSocket, but we define it here
  # so clients know it exists. The payload types are in components.
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: integer }
        details: { type: string }
    StatusResponse:
      type: object
      required: [status, version, uptime, hostname, ip_address, port, secret]
      properties:
        status: { type: string }
        version: { type: string }
        uptime: { type: string }
        hostname: { type: string, format: hostname }
        ip_address:
          oneOf:
            - type: string
              format: ipv4
            - type: string
              format: ipv6
        port: { type: string, format: uint16 }
        secret: { type: string }
    Monitor:
      type: object
      required: [edid, port, name, manufacturer]
      properties:
        edid: { type: string }
        port: { type: string }
        name: { type: string }
        manufacturer: { type: string }
        active: { $ref: '#/components/schemas/ActiveMonitor' }
        # Registry-derived control metadata (optional).
        friendly_name: { type: string }
        control_backend: { type: string, description: "ddc | tv | none" }
        brightness: { type: integer, description: "current 0-100, -1 if unknown" }
        capabilities: { $ref: '#/components/schemas/MonitorCapabilities' }
        visibility:
          type: object
          additionalProperties: { type: boolean }
    MonitorCapabilities:
      type: object
      required: [brightness, power, volume]
      properties:
        brightness: { type: boolean }
        power: { type: boolean }
        volume: { type: boolean }
    ActiveMonitor:
      type: object
      required: [width, height, refresh_rate, position_x, position_y, primary, model]
      properties:
        width: { type: integer }
        height: { type: integer }
        refresh_rate: { type: number, format: double }
        position_x: { type: integer }
        position_y: { type: integer }
        primary: { type: boolean }
        model: { type: string }
    MonitorsResponse:
      type: array
      items:
        $ref: '#/components/schemas/Monitor'

    # --- Layout Types ---
    LayoutMonitor:
      type: object
      required: [name, edid, port, width, height, refresh_rate, position_x, position_y, primary]
      properties:
        name: { type: string }
        edid: { type: string }
        port: { type: string }
        width: { type: integer }
        height: { type: integer }
        refresh_rate: { type: number, format: double }
        position_x: { type: integer }
        position_y: { type: integer }
        primary: { type: boolean }
    Layout:
      type: object
      required: [id, name, monitors, aliases]
      properties:
        id: { type: string }
        name: { type: string }
        emoji: { type: string }
        aliases:
          type: array
          items: { type: string }
        monitors:
          type: array
          items:
            $ref: '#/components/schemas/LayoutMonitor'
    LayoutsResponse:
      type: object
      required: [layouts, current_layout]
      properties:
        layouts:
          type: array
          items:
            $ref: '#/components/schemas/Layout'
        current_layout: { type: string }
    SwitchLayoutRequest:
      type: object
      required: [layout]
      properties:
        layout: { type: string }
    SwitchLayoutResponse:
      type: object
      required: [success, current_layout]
      properties:
        success: { type: boolean }
        current_layout: { type: string }
        message: { type: string }
...
    SaveLayoutRequest:
      type: object
      required: [name]
      properties:
        id: { type: string }
        name: { type: string }
        emoji: { type: string }
    SaveLayoutResponse:
      type: object
      required: [success, layout]
      properties:
        success: { type: boolean }
        layout: { $ref: '#/components/schemas/Layout' }
        message: { type: string }
    RemoveLayoutRequest:
      type: object
      required: [layout]
      properties:
        layout: { type: string }
    RemoveLayoutResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }
    ShutdownResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }

    # --- Audio Types ---
...
    SimStateResponse:
      type: object
      required: [state]
      properties:
        state: { type: string }
    SimSetStateRequest:
      type: object
      required: [state]
      properties:
        state: { type: string }

    # --- WebSocket Types ---
...